Give browser page helpers a core owner
Move page diagnostics, readiness probing, safe title lookup, and smart navigation into src/core/page so browser Page primitives are not mixed into the generic utility bucket.

Constraint: The cleanup is being landed as narrow ownership PRs while preserving public tool behavior.

Rejected: Keep browser Page helpers in src/utils | They are shared domain primitives used by session, CDP, Chrome fallback, and tools.

Confidence: high

Scope-risk: moderate

Directive: New browser Page lifecycle or diagnosis helpers should be added under src/core/page unless they belong to a single tool.

Tested: npm test -- --runTestsByPath tests/core/page/diagnostics.test.ts tests/core/page/ready-state.test.ts tests/core/page/smart-goto.test.ts tests/core/page/safe-title.test.ts tests/captcha/captcha-detection.test.ts tests/captcha/captcha-handler.test.ts tests/dom/shadow-dom-utils.test.ts tests/tools/extract-data-ready.test.ts tests/tools/navigate.test.ts tests/tools/navigate-auto-fallback.test.ts tests/tools/navigate-headed-fallback.test.ts tests/tools/navigate-stealth.test.ts tests/tools/navigate.auto-recall.test.ts tests/tools/validate-page.test.ts tests/cdp/client-contracts.test.ts tests/cdp/create-page-startup-reuse.test.ts tests/headed/headed-fallback-manager.test.ts --runInBand; npm run build; npm run lint; npm run lint:tier; npm run lint:repo-structure; git diff --check

Not-tested: Full manual workflow_dispatch test matrix.
