+++ src/chrome/headed-fallback.ts
@@ -17,8 +17,8 @@ import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
 import { hasDisplay } from '../utils/display-detect';
-import { detectBlockingPage, BlockingInfo } from '../utils/page-diagnostics';
-import { safeTitle } from '../utils/safe-title';
+import { detectBlockingPage, BlockingInfo } from '../core/page/diagnostics';
+import { safeTitle } from '../core/page/safe-title';
 import { getTargetId } from '../utils/puppeteer-helpers';
 import { spawnProcessGuardian } from './process-guardian';
 import { writeMarker, removeMarker } from './ownership-marker';
