+++ src/core/page/ready-state.ts
@@ -1,8 +1,8 @@
 import type { Page } from 'puppeteer-core';
-import type { ToolContext } from '../types/mcp';
-import { getRemainingBudget } from '../types/mcp';
-import { isTimeoutError } from '../errors/timeout';
-import { withTimeout } from './with-timeout';
+import type { ToolContext } from '../../types/mcp';
+import { getRemainingBudget } from '../../types/mcp';
+import { isTimeoutError } from '../../errors/timeout';
+import { withTimeout } from '../../utils/with-timeout';
 
 export interface PageReadyOptions {
   /** Maximum wall-clock wait in ms. Default: 5000. */
