+++ src/tools/session-snapshot.ts
@@ -13,7 +13,7 @@ import { getSessionManager } from '../session-manager';
 import { getExistingChromeLauncher } from '../chrome/launcher';
 import { getGlobalConfig } from '../config/global';
 import { writeFileAtomicSafe } from '../utils/atomic-file';
-import { safeTitle } from '../utils/safe-title';
+import { safeTitle } from '../core/page/safe-title';
 import { currentRequestContext } from '../observability/request-id';
 import { assertFilePathAllowedBySessionRoots } from '../security/mcp-roots';
