+++ src/tools/validate-page.ts
@@ -16,8 +16,8 @@ import { MCPServer } from '../mcp-server';
 import { MCPToolDefinition, MCPResult, ToolHandler } from '../types/mcp';
 import { TOOL_ANNOTATIONS } from '../types/tool-annotations';
 import { getSessionManager } from '../session-manager';
-import { smartGoto } from '../utils/smart-goto';
-import { safeTitle } from '../utils/safe-title';
+import { smartGoto } from '../core/page/smart-goto';
+import { safeTitle } from '../core/page/safe-title';
 import { assertDomainAllowed } from '../security/domain-guard';
 import { buildTextMetrics } from '../core/metrics/token-estimate';
 import { isStateHeaderEnabled, prependHeaderText } from './_shared/state-header';
