+++ tests/cdp/create-page-startup-reuse.test.ts
@@ -24,7 +24,7 @@ jest.mock('../../src/config/global', () => ({
 }));
 
 const smartGoto = jest.fn();
-jest.mock('../../src/utils/smart-goto', () => ({
+jest.mock('../../src/core/page/smart-goto', () => ({
   smartGoto: (...args: unknown[]) => smartGoto(...args),
 }));
