+++ tests/tools/navigate-auto-fallback.test.ts
@@ -17,20 +17,20 @@ jest.mock('../../src/session-manager', () => ({
 }));
 
 // Mock smart-goto
-import type { SmartGotoResult } from '../../src/utils/smart-goto';
+import type { SmartGotoResult } from '../../src/core/page/smart-goto';
 const mockSmartGotoFn = jest.fn<Promise<SmartGotoResult>, [any, string, any?]>(
   async (page, url, opts) => {
     await page.goto(url, opts);
     return { response: null };
   },
 );
-jest.mock('../../src/utils/smart-goto', () => ({
+jest.mock('../../src/core/page/smart-goto', () => ({
   smartGoto: mockSmartGotoFn,
 }));
 
 // Mock page-diagnostics to control blocking detection
 const mockDetectBlockingPage = jest.fn().mockResolvedValue(null);
-jest.mock('../../src/utils/page-diagnostics', () => ({
+jest.mock('../../src/core/page/diagnostics', () => ({
   detectBlockingPage: (...args: any[]) => mockDetectBlockingPage(...args),
   BlockingInfo: {},
 }));
@@ -72,10 +72,10 @@ describe('NavigateTool - Auto-fallback (#459)', () => {
     jest.doMock('../../src/session-manager', () => ({
       getSessionManager: () => mockSessionManager,
     }));
-    jest.doMock('../../src/utils/smart-goto', () => ({
+    jest.doMock('../../src/core/page/smart-goto', () => ({
       smartGoto: mockSmartGotoFn,
     }));
-    jest.doMock('../../src/utils/page-diagnostics', () => ({
+    jest.doMock('../../src/core/page/diagnostics', () => ({
       detectBlockingPage: (...args: any[]) => mockDetectBlockingPage(...args),
       BlockingInfo: {},
     }));
