+++ tests/tools/navigate-stealth.test.ts
@@ -13,7 +13,7 @@ jest.mock('../../src/session-manager', () => ({
 }));
 
 // Mock smart-goto
-import type { SmartGotoResult } from '../../src/utils/smart-goto';
+import type { SmartGotoResult } from '../../src/core/page/smart-goto';
 
 const mockSmartGotoFn = jest.fn<Promise<SmartGotoResult>, [any, string, any?]>(
   async (page, url, opts) => {
@@ -21,7 +21,7 @@ const mockSmartGotoFn = jest.fn<Promise<SmartGotoResult>, [any, string, any?]>(
     return { response: null };
   },
 );
-jest.mock('../../src/utils/smart-goto', () => ({
+jest.mock('../../src/core/page/smart-goto', () => ({
   smartGoto: mockSmartGotoFn,
 }));
 
@@ -36,7 +36,7 @@ describe('NavigateTool - Stealth Mode', () => {
     jest.doMock('../../src/session-manager', () => ({
       getSessionManager: () => mockSessionManager,
     }));
-    jest.doMock('../../src/utils/smart-goto', () => ({
+    jest.doMock('../../src/core/page/smart-goto', () => ({
       smartGoto: mockSmartGotoFn,
     }));
     const { registerNavigateTool } = await import('../../src/tools/navigate');
